Improve typing of language and auth
