+++ app/api/files/specs/routes.spec.ts
@@ -15,7 +15,7 @@ import { files } from '../files';
 import uploadRoutes from '../routes';
 
 jest.mock(
-  '../../auth/authMiddleware.js',
+  '../../auth/authMiddleware.ts',
   () => () => (_req: Request, _res: Response, next: NextFunction) => {
     next();
   }
