+++ app/api/sync/specs/uploadRoute.spec.js
@@ -9,7 +9,7 @@ import fs from 'fs';
 import syncRoutes from '../routes.js';
 import paths from '../../config/paths';
 
-jest.mock('../../auth/authMiddleware.js', () => () => (_req, _res, next) => {
+jest.mock('../../auth/authMiddleware.ts', () => () => (_req, _res, next) => {
   next();
 });
