+++ app/api/tasks/specs/routes.spec.ts
@@ -5,7 +5,7 @@ import request from 'supertest';
 import testRoute from '../routes';
 
 jest.mock(
-  '../../auth/authMiddleware.js',
+  '../../auth/authMiddleware.ts',
   () => () => (_req: Request, _res: Response, next: NextFunction) => {
     next();
   }
